+++ Components/Host.php
@@ -22,6 +22,7 @@ use League\Uri\Exceptions\SyntaxError;
 use League\Uri\Idna\Converter as IdnConverter;
 use League\Uri\IPv4\Converter as IPv4Converter;
 use League\Uri\IPv4Normalizer;
+use League\Uri\IPv6\Converter;
 use Psr\Http\Message\UriInterface as Psr7UriInterface;
 use Stringable;